+++ src/main/java/com/maxmind/geoip2/record/City.java
@@ -10,13 +10,9 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This record is returned by all the end points except the Country end point.
  * </p>
  */
-final public class City extends AbstractNamedRecord {
+public final class City extends AbstractNamedRecord {
     @JsonProperty
     private Integer confidence;
-
-    public City() {
-        super();
-    }
 
     /**
      * @return A value from 0-100 indicating MaxMind's confidence that the city