+++ src/main/java/com/maxmind/geoip2/record/Subdivision.java
@@ -10,15 +10,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This record is returned by all the end points except the Country end point.
  * </p>
  */
-final public class Subdivision extends AbstractNamedRecord {
+public final class Subdivision extends AbstractNamedRecord {
     @JsonProperty
     private Integer confidence;
 
     @JsonProperty("iso_code")
     private String isoCode;
-
-    public Subdivision() {
-    }
 
     /**
      * @return This is a value from 0-100 indicating MaxMind's confidence that